Extracted ValidatorHistoricPerformance to separate component
