+++ packages/page-staking/src/Query/AlephCommitttee/README.md
@@ -1,0 +1,3 @@
+# Overview
+
+This folder contains logic specific for AlephZero chain.
